Continue work on AREQ-24774 -- Import REQ Database from BXT IVI O, O MR1 to KBL-POS
In process. This still needs code to output new Feature and E-Features.

Spent time troubleshooting query--when source=target, all records should already be in place, but it was finding an [O-MR1] e-feature with and saying that wasn't the same as what it should have been scanning. Problem was that the summary had been changed but the version in the record reflected O--So when [O][Broxton...] was scanned via query it pulled up this record with O in the version but O-MR1 in the summary! :-\

Change-Id: Ia223651854ec195bcc8d78dc23beebcd6ce98893